+++ src/Kinematics.h
@@ -5,5 +5,5 @@
 #define FOREARM_LENGTH 220.0
 #define BICEPS_LENGTH 140.0
 
-int delta_calcForward(float theta1, float theta2, float theta3, float &x0, float &y0, float &z0);
-int delta_calcInverse(float x0, float y0, float z0, float &theta1, float &theta2, float &theta3);
+int delta_calcForward(double theta1, double theta2, double theta3, double &x0, double &y0, double &z0);
+int delta_calcInverse(double x0, double y0, double z0, double &theta1, double &theta2, double &theta3);
